addReply + eval commands added
