+++ index.js
@@ -7,7 +7,8 @@ const path = require("path");
 require('dotenv').config();
 
 const app = express();
-const bot = new Discord.Client({ intents: ["GUILDS", "GUILD_MESSAGES", "GUILD_VOICE_STATES"], disableMentions: 'everyone' });
+const bot = new Discord.Client({  intents: ["GUILDS", "GUILD_MESSAGES", "GUILD_VOICE_STATES", "DIRECT_MESSAGES"], 
+                                  disableMentions: 'everyone' });
 
 const text = fs.readFileSync(path.join(__dirname, "./assets/replyArray.txt"), "utf-8");
 const replyArray = text.split("\n");
